Id b 14 pet creation

+++ Backend/src/PetFamily.Application/Volunteers/Delete/DeleteVolunteerHandler.cs
@@ -1,5 +1,8 @@
 ﻿using CSharpFunctionalExtensions;
+using FluentValidation;
 using Microsoft.Extensions.Logging;
+using PetFamily.Application.DataBase;
+using PetFamily.Application.Extensions;
 using PetFamily.Application.Interfaces;
 using PetFamily.Domain.Shared;
 
@@ -9,22 +12,33 @@ public class DeleteVolunteerHandler
 {
     private readonly IVolunteerRepository _volunteerRepository;
     private readonly ILogger<DeleteVolunteerHandler> _logger;
+    private readonly IUnitOfWork _unitOfWork;
+    private readonly IValidator<DeleteVolunteerCommand> _validator;
 
     public DeleteVolunteerHandler(IVolunteerRepository volunteerRepository,
-        ILogger<DeleteVolunteerHandler> logger)
+        ILogger<DeleteVolunteerHandler> logger,
+        IUnitOfWork unitOfWork, IValidator<DeleteVolunteerCommand> validator)
     {
         _volunteerRepository = volunteerRepository;
         _logger = logger;
+        _unitOfWork = unitOfWork;
+        _validator = validator;
     }
     
-    public async Task<Result<Guid, CustomError>> Handle(DeleteVolunteerRequest request,
+    public async Task<Result<Guid, CustomErrorsList>> Handle(DeleteVolunteerCommand command,
         CancellationToken cancellationToken = default)
     {
-        var volunteerResult = await _volunteerRepository.GetById(request.VolunteerId, cancellationToken);
+        var validationResult = _validator.Validate(command);
+        if (!validationResult.IsValid)
+            return validationResult.ToErrorList();
+        
+        var volunteerResult = await _volunteerRepository.GetById(command.VolunteerId, cancellationToken);
         if (volunteerResult.IsFailure)
-            return volunteerResult.Error;
+            return volunteerResult.Error.ToErrorList();
         
         var result = await _volunteerRepository.Delete(volunteerResult.Value, cancellationToken);
+
+        await _unitOfWork.SaveChanges(cancellationToken);
         
         _logger.LogInformation("Volunteer with Id: {id} was deleted", volunteerResult.Value.Id);
